Renamed interface and added better random int generator
Renamed interface so that it's clear that we're talking about int number generators. Added a new class based on the RNGCryptoServiceProvider. This is a much better generator, but it will only return arrays. In practice, this means we'll have to write more code in order to get the ints we require.